+++ svc/sdr.zsh
@@ -1,6 +1,6 @@
 #!/bin/zsh
-# $Id: mkinitramfs-ll/svc/sdr.zsh,v 0.11.2 2012/11/05 21:11:36 -tclover Exp $
-revision=0.11.2
+# $Id: mkinitramfs-ll/svc/sdr.zsh,v 0.11.7 2013/02/11 09:59:45 -tclover Exp $
+revision=0.11.7
 usage() {
   cat <<-EOF
  usage: ${(%):-%1x} [-update|-remove] [-r|-sqfsdir<dir>] -d|-sqfsd:<dir>:<dir>
@@ -43,11 +43,39 @@ if [[ -n $(uname -m | grep 64) ]] { opts[-arc]=64 } else { opts[-arc]=32 }
 :	${opts[-busybox]:=${opts[-B]:-$(which bb)}}
 info() 	{ print -P " %B%F{green}*%b%f $@" }
 error() { print -P " %B%F{red}*%b%f $@" }
-die()   { error $@; return 1 }
+die()   { error $@; return }
 alias die='die "%F{yellow}%1x:%U${(%):-%I}%u:%f" $@'
 setopt NULL_GLOB
+mnt() {
+	if [[ $dir = *bin ]] || [[ $dir = lib* ]] {
+		local busybox=/tmp/busybox cp grep mount mv rm mcdir mrc
+		cp ${opts[-busybox]} /tmp/busybox ||
+			die "no static busybox binary found"
+		cp="$busybox cp -ar"; mv="$busybox mv"; rm="$busybox rm -fr"
+		mount="$busybox mount"; umount="$busybox umount"; grep="$busybox grep"
+	} else { cp="cp -ar"; grep=grep; mount=mount; umount=umount; mv=mv; rm="rm -fr" }
+	if [[ -n $(${=mount} -t aufs | ${=grep} -w $dir) ]] {
+		${=umount} -l /$dir 1>/dev/null 2>&1 || die "$dir: failed to umount aufs branch"
+	}
+	if [[ -n $(${=mount} -t squashfs | ${=grep} $bdir/rr) ]] {
+		${=umount} -l $bdir/rr 1>/dev/null 2>&1 || die "failed to umount $bdir.sfs" 
+	}
+	${=rm} $bdir/rw/* || die "failed to clean up $bdir/rw"
+	[[ -f $bdir.sfs ]] && ${=rm} $bdir.sfs 
+	${=mv} $bdir.tmp.sfs $bdir.sfs || die "failed to move $dir.tmp.sfs"
+	${=mount} $bdir.sfs $bdir/rr -t squashfs -o nodev,loop,ro 1>/dev/null 2>&1 &&
+	{	
+		if [[ -n ${(k)opts[-R]} ]] || [[ -n ${(k)opts[-remove]} ]] { 
+			${=rm} /$dir/* || die "failed to clean up $bdir"
+		} 
+		if [[ -n ${(k)opts[-U]} ]] || [[ -n ${(k)opts[-update]} ]] { 
+			${=rm} /$dir && ${=cp} $bdir/rr /$dir || die "$dir: failed to update"
+		}
+	${=mount} -t aufs -o nodev,udba=reval,br:$bdir/rw:$bdir/rr $dir /$dir \
+		1>/dev/null 2>&1 || die "$dir: failed to mount aufs branch"
+	} || die "failed to mount $dir.sfs"
+}
 squashd() {
-	local cp mv rm mcdir mrc
 	if [[ -n ${(k)opts[-fstab]} || -n ${(k)opts[-fstab]} ]] {
 		echo "$bdir.sfs $bdir/rr squashfs nodev,loop,rr 0 0" >>/etc/fstab ||
 			die "$dir: failed to write squasshfs line"
@@ -69,33 +97,8 @@ squashd() {
 				mrc=yes || die "failed to move rc-svcdir"
 		}
 	}
-	if [[ -n $(mount -t aufs | grep -w $dir) ]] {
-		umount -l /$dir 1>/dev/null 2>&1 || die "$dir: failed to umount aufs branch"
-	}
-	if [[ -n $(mount -t squashfs | grep $bdir/rr) ]] {
-		umount -l $bdir/rr 1>/dev/null 2>&1 || die "failed to umount $bdir.sfs" 
-	}
-	if [[ $dir = *bin ]] || [[ $dir = lib* ]] {
-		busybox=/tmp/busybox; cp ${opts[-busybox]} /tmp/busybox ||
-			die "no static busybox binary found"
-		cp="$busybox cp -ar"; mv="$busybox mv"; rm="$busybox rm -fr"
-	} else { cp="cp -ar"; mv=mv; rm="rm -fr" }
-	${=rm} $bdir/rw/* || die "failed to clean up $bdir/rw"
-	[[ -e $bdir.sfs ]] && ${=rm} $bdir.sfs 
-	${=mv} $bdir.tmp.sfs $bdir.sfs || die "failed to move $dir.tmp.sfs"
-	if [[ -n ${(k)opts[-n]} ]] || [[ -n ${(k)opts[-nomount]} ]] { :; } else {
-		mount $bdir.sfs $bdir/rr -tsquashfs -onodev,loop,ro 1>/dev/null 2>&1 &&
-		{	
-		if [[ -n ${(k)opts[-R]} ]] || [[ -n ${(k)opts[-remove]} ]] { 
-			${=rm} /$dir/* || die "failed to clean up $bdir"
-		} 
-		if [[ -n ${(k)opts[-U]} ]] || [[ -n ${(k)opts[-update]} ]] { 
-			${=rm} /$dir && ${=cp} $bdir/rr /$dir || die "$dir: failed to update"
-		}
-		mount -onodev,udba=reval,br:$bdir/rw:$bdir/rr -taufs $dir /$dir 1>/dev/null 2>&1 ||
-			die "$dir: failed to mount aufs branch"
-		} || die "failed to mount $dir.sfs"
-	}
+	if [[ -n ${(k)opts[-n]} ]] || [[ -n ${(k)opts[-nomount]} ]] { :;
+	} else { mnt }
 	if [[ -n $mcdir ]] { 
 		mount --move /var/cache/splash /$dir/splash/cache 1>/dev/nul 2>&1 ||
 			die "failed to move back cachedir"
@@ -118,6 +121,6 @@ for dir (${(pws,:,)opts[-sqfsd]} ${(pws,:,)opts[-d]}) {
 		} else { print -P "%F{green}>>> updating squashed $dir...%f"; squashd }
 	} else { print -P "%F{green}>>> building squashed $dir...%f"; squashd }
 }
-rm -f $busybox
+[[ -f $busybox ]] && rm -f $busybox
 unset bdir opts rr rw
 # vim:fenc=utf-8:ci:pi:sts=0:sw=4:ts=4:
